Add files via upload
A simple brute force approach would be iterate through the array and for each index i in the array. Next loop will iterate through the array once again to check if there are any matching values to create a good pair. If there are, we increment declared "cont" variable by 1.